requirements.txt
docker~=5.0.2
blessed==1.19.0
pydantic~=1.8.2
python-dotenv~=0.19.0
cursor~=1.3.4
urwid~=2.1.2
rich~=10.12.0
ascii-graph~=1.5.1
